+++ .vscode/settings.json
@@ -1,33 +1,1 @@
-{
-  "editor.tabSize": 4,
-  "editor.insertSpaces": true,
-  "files.eol": "\n",
-  "files.watcherExclude": {
-    "**/.git/objects/**": true,
-    "**/node_modules/**": true,
-    ".tmp": true
-  },
-  "files.exclude": {
-    ".tmp": true,
-    "coverage": true
-  },
-  "search.exclude": {
-    ".tmp": true,
-    "dist": true,
-    "coverage": true
-  },
-  "json.schemas": [
-    {
-      "fileMatch": [
-        "/pbiviz.json"
-      ],
-      "url": "./.api/v1.13.0/schema.pbiviz.json"
-    },
-    {
-      "fileMatch": [
-        "/capabilities.json"
-      ],
-      "url": "./.api/v1.13.0/schema.capabilities.json"
-    }
-  ]
-}
+{"editor.tabSize":4,"editor.insertSpaces":true,"files.eol":"\n","files.watcherExclude":{"**/.git/objects/**":true,"**/node_modules/**":true,".tmp":true},"files.exclude":{".tmp":true},"search.exclude":{".tmp":true,"typings":true},"json.schemas":[{"fileMatch":["/pbiviz.json"],"url":"./.api/v2.1.0/schema.pbiviz.json"},{"fileMatch":["/capabilities.json"],"url":"./.api/v2.1.0/schema.capabilities.json"},{"fileMatch":["/dependencies.json"],"url":"./.api/v2.1.0/schema.dependencies.json"}]}
