+++ karma.conf.ts
@@ -1,0 +1,134 @@
+/*
+ *  Power BI Visualizations
+ *
+ *  Copyright (c) Microsoft Corporation
+ *  All rights reserved.
+ *  MIT License
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the ""Software""), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in
+ *  all copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ *  THE SOFTWARE.
+ */
+
+"use strict";
+
+const webpackConfig = require("./test.webpack.config.js");
+const tsconfig = require("./test.tsconfig.json");
+const path = require("path");
+
+const testRecursivePath = "test/visualTest.ts";
+const srcOriginalRecursivePath = "src/**/*.ts";
+const srcRecursivePath = ".tmp/drop/**/*.js";
+const coverageFolder = "coverage";
+const globals = "./test/globals.ts";
+
+process.env.CHROME_BIN = require("puppeteer").executablePath();
+
+import { Config, ConfigOptions } from "karma";
+
+module.exports = (config: Config) => {
+    config.set(<ConfigOptions>{
+        browserNoActivityTimeout: 100000,
+        browsers: ["ChromeHeadless"],
+        colors: true,
+        frameworks: ["jasmine"],
+        reporters: [
+            "progress",
+            "coverage",
+            "coverage-istanbul"
+        ],
+        singleRun: true,
+        plugins: [
+            "karma-coverage",
+            "karma-typescript",
+            "karma-webpack",
+            "karma-jasmine",
+            "karma-sourcemap-loader",
+            "karma-chrome-launcher",
+            "karma-coverage-istanbul-reporter"
+        ],
+        files: [
+            "node_modules/jquery/dist/jquery.min.js",
+            "node_modules/jasmine-jquery/lib/jasmine-jquery.js",
+            globals,
+            srcRecursivePath,
+            testRecursivePath,
+            {
+                pattern: srcOriginalRecursivePath,
+                included: false,
+                served: true
+            },
+            {
+                pattern: "./capabilities.json",
+                watched: false,
+                served: true,
+                included: false
+            }
+        ],
+        preprocessors: {
+            [testRecursivePath]: ["webpack"],
+            [srcRecursivePath]: ["sourcemap"]
+        },
+        typescriptPreprocessor: {
+            options: tsconfig.compilerOptions
+        },
+        coverageIstanbulReporter: {
+            // reports can be any that are listed here: https://github.com/istanbuljs/istanbuljs/tree/aae256fb8b9a3d19414dcf069c592e88712c32c6/packages/istanbul-reports/lib
+            reports: {
+                lcovonly: coverageFolder + "lcov.info",
+                html: coverageFolder,
+                "text-summary": null
+            },
+
+            // base output directory. If you include %browser% in the path it will be replaced with the karma browser name
+            dir: path.join(__dirname, "coverage"),
+
+            // Combines coverage information from multiple browsers into one report rather than outputting a report
+            // for each browser.
+            combineBrowserReports: true,
+
+            // if using webpack and pre-loaders, work around webpack breaking the source path
+            fixWebpackSourcePaths: true,
+
+            // Omit files with no statements, no functions and no branches from the report
+            skipFilesWithNoCoverage: true,
+
+            // Most reporters accept additional config options. You can pass these through the `report-config` option
+            "report-config": {
+                // all options available at: https://github.com/istanbuljs/istanbuljs/blob/aae256fb8b9a3d19414dcf069c592e88712c32c6/packages/istanbul-reports/lib/html/index.js#L135-L137
+                html: {
+                    // outputs the report in ./coverage/html
+                    subdir: "html"
+                }
+            }
+        },
+        coverageReporter: {
+            dir: coverageFolder,
+            reporters: [
+                { type: "html" },
+                { type: "lcov" }
+            ]
+        },
+        mime: {
+            "text/x-typescript": ["ts", "tsx"]
+        },
+        webpack: webpackConfig,
+        webpackMiddleware: {
+            stats: "errors-only"
+        }
+    });
+};
