+++ package.json
@@ -1,17 +1,19 @@
 {
   "name": "powerbi-visuals-sunburst",
   "description": "SunBurst chart.",
-  "version": "1.5.0",
+  "version": "2.0.0",
   "author": {
     "name": "Microsoft",
     "email": "pbicvsupport@microsoft.com"
   },
   "scripts": {
-    "postinstall": "pbiviz update 1.13.0",
     "pbiviz": "pbiviz",
     "start": "pbiviz start",
     "package": "pbiviz package",
-    "lint": "tslint -r \"node_modules/tslint-microsoft-contrib\" \"+(src|test)/**/*.ts\"",
+    "cert": "pbiviz --create-cert",
+    "lint": "npm run lint:ts && npm run lint:less",
+    "lint:ts": "node node_modules/tslint/bin/tslint \"+(src|test)/**/*.ts\"",
+    "lint:less": "stylelint -s less \"style/**/*.less\"",
     "pretest": "pbiviz package --resources --no-minify --no-pbiviz --no-plugin",
     "test": "karma start"
   },
@@ -20,40 +22,52 @@
     "type": "git",
     "url": "git+https://github.com/Microsoft/PowerBI-visuals-sunburst.git"
   },
-  "dependencies": {
-    "d3": "3.5.5",
-    "lodash": "4.17.4",
-    "powerbi-visuals-utils-dataviewutils": "^1.2.0",
-    "powerbi-visuals-utils-svgutils": "^1.1.0",
-    "powerbi-visuals-utils-tooltiputils": "1.0.0",
-    "powerbi-visuals-utils-formattingutils": "^2.1.0",
-    "powerbi-visuals-utils-typeutils": "^1.1.0",
-    "powerbi-visuals-utils-colorutils": "^1.1.0",
-    "powerbi-visuals-utils-chartutils": "^1.2.0",
-    "powerbi-visuals-utils-interactivityutils": "3.1.0",
-    "powerbi-models": "1.0.3"
-  },
   "devDependencies": {
-    "@types/d3": "3.5.36",
-    "@types/jasmine": "2.5.53",
-    "@types/jasmine-jquery": "1.5.31",
-    "@types/jquery": "2.0.41",
+    "@types/d3": "^5.0.0",
+    "@types/jasmine": "^2.5.53",
+    "@types/jasmine-jquery": "^1.5.31",
+    "@types/jquery": "2.0.45",
+    "@types/karma": "1.7.3",
     "@types/lodash": "4.14.55",
+    "@types/webpack": "4.4.4",
     "coveralls": "3.0.2",
-    "jasmine": "2.6.0",
+    "css-loader": "1.0.0",
+    "jasmine": "2.5.2",
+    "jasmine-jquery": "2.1.1",
     "jquery": "3.2.1",
-    "karma": "2.0.4",
-    "karma-chrome-launcher": "2.2.0",
+    "karma": "3.0.0",
+    "karma-chrome-launcher": "^2.2.0",
     "karma-coverage": "1.1.2",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
     "karma-jasmine": "1.1.2",
-    "karma-remap-istanbul": "0.6.0",
     "karma-sourcemap-loader": "0.3.7",
+    "karma-typescript": "3.0.13",
     "karma-typescript-preprocessor": "0.3.1",
-    "powerbi-visuals-tools": "1.13.1",
-    "powerbi-visuals-utils-testutils": "1.2.1",
-    "puppeteer": "^1.6.0",
-    "tslint": "5.4.3",
+    "karma-webpack": "3.0.0",
+    "less": "^2.1.1",
+    "less-loader": "^4.1.0",
+    "powerbi-visuals-api": "^2.1.0",
+    "powerbi-visuals-tools": "3.0.2",
+    "powerbi-visuals-utils-testutils": "^2.0.4",
+    "puppeteer": "^1.6.1",
+    "style-loader": "0.21.0",
+    "stylelint": "^7.12.0",
+    "ts-loader": "4.4.2",
+    "ts-node": "^7.0.0",
+    "tslint": "5.11.0",
+    "webpack": "4.14.0",
     "tslint-microsoft-contrib": "^5.0.0",
-    "typescript": "2.3.4"
+    "typescript": "^3.0.1",
+    "d3": "^5.7.0",
+    "lodash": "4.17.4",
+    "powerbi-models": "^1.0.8",
+    "powerbi-visuals-utils-chartutils": "2.0.7",
+    "powerbi-visuals-utils-colorutils": "2.0.2",
+    "powerbi-visuals-utils-dataviewutils": "2.0.1",
+    "powerbi-visuals-utils-formattingutils": "4.0.2",
+    "powerbi-visuals-utils-interactivityutils": "^5.0.1",
+    "powerbi-visuals-utils-svgutils": "2.0.2",
+    "powerbi-visuals-utils-tooltiputils": "2.0.3",
+    "powerbi-visuals-utils-typeutils": "2.0.1"
   }
 }
