+++ src/behavior.ts
@@ -24,128 +24,135 @@
  *  THE SOFTWARE.
  */
 
-module powerbi.extensibility.visual.behavior {
-    import ISelectionHandler = powerbi.extensibility.utils.interactivity.ISelectionHandler;
-    import IInteractiveBehavior = powerbi.extensibility.utils.interactivity.IInteractiveBehavior;
-    import IInteractivityService = powerbi.extensibility.utils.interactivity.IInteractivityService;
-    import InteractivityServiceBase = powerbi.extensibility.utils.interactivity.InteractivityService;
+import { Selection, event as d3Event } from "d3";
 
-    export const DimmedOpacity: number = 0.2;
-    export const DefaultOpacity: number = 1.0;
+import powerbi from "powerbi-visuals-api";
+import IVisualHost = powerbi.extensibility.visual.IVisualHost;
+import ISelectionId = powerbi.visuals.ISelectionId;
 
-    export function getFillOpacity(
-        selected: boolean,
-        highlight: boolean,
-        hasSelection: boolean,
-        hasPartialHighlights: boolean
-    ): number {
-        if ((hasPartialHighlights && !highlight) || (hasSelection && !selected)) {
-            return DimmedOpacity;
-        }
+import { interactivityService } from "powerbi-visuals-utils-interactivityutils";
+import IInteractiveBehavior = interactivityService.IInteractiveBehavior;
+import IInteractivityService = interactivityService.IInteractivityService;
+import InteractivityServiceBase = interactivityService.InteractivityService;
+import ISelectionHandler = interactivityService.ISelectionHandler;
 
-        return DefaultOpacity;
+import { SunburstDataPoint } from "./dataInterfaces";
+
+const DimmedOpacity: number = 0.2;
+const DefaultOpacity: number = 1.0;
+
+function getFillOpacity(
+    selected: boolean,
+    highlight: boolean,
+    hasSelection: boolean,
+    hasPartialHighlights: boolean
+): number {
+    if ((hasPartialHighlights && !highlight) || (hasSelection && !selected)) {
+        return DimmedOpacity;
     }
 
-    export interface BehaviorOptions {
-        dataPoints: SunburstDataPoint[];
-        selection: d3.Selection<SunburstDataPoint>;
-        clearCatcher: d3.Selection<any>;
-        interactivityService: IInteractivityService;
-        onSelect?: (dataPoint: SunburstDataPoint) => void;
-    }
+    return DefaultOpacity;
+}
 
-    export class Behavior implements IInteractiveBehavior {
-        private options: BehaviorOptions;
+export interface BehaviorOptions {
+    dataPoints: SunburstDataPoint[];
+    selection: Selection< d3.BaseType, SunburstDataPoint, d3.BaseType, SunburstDataPoint>;
+    clearCatcher: Selection<d3.BaseType, any, d3.BaseType, any>;
+    interactivityService: IInteractivityService;
+    onSelect?: (dataPoint: SunburstDataPoint) => void;
+}
 
-        public bindEvents(
-            options: BehaviorOptions,
-            selectionHandler: ISelectionHandler
-        ): void {
-            this.options = options;
+export class Behavior implements IInteractiveBehavior {
+    private options: BehaviorOptions;
 
-            const {
-                selection,
-                clearCatcher,
-                onSelect,
-            } = options;
+    public bindEvents(
+        options: BehaviorOptions,
+        selectionHandler: ISelectionHandler
+    ): void {
+        this.options = options;
 
-            selection.on("click", (dataPoint: SunburstDataPoint) => {
-                const event: Event = d3.event as Event;
+        const {
+            selection,
+            clearCatcher,
+            onSelect,
+        } = options;
 
-                selectionHandler.handleSelection(dataPoint, false);
+        selection.on("click", (dataPoint: SunburstDataPoint) => {
+            const event: Event = d3Event as Event;
 
-                event.stopPropagation();
+            selectionHandler.handleSelection(dataPoint, false);
 
-                if (onSelect) {
-                    onSelect(dataPoint);
-                }
-            });
+            event.stopPropagation();
 
-            clearCatcher.on("click", () => {
-                selectionHandler.handleClearSelection();
-
-                if (onSelect) {
-                    onSelect(null);
-                }
-            });
-        }
-
-        public renderSelection(hasSelection: boolean): void {
-            const {
-                selection,
-                interactivityService,
-            } = this.options;
-
-            this.options.dataPoints
-                .filter((dataPoint: SunburstDataPoint) => dataPoint && dataPoint.selected)
-                .forEach(this.markDataPointsAsSelected.bind(this));
-
-            const hasHighlights: boolean = interactivityService.hasSelection();
-
-            selection.style("opacity", (dataPoint: SunburstDataPoint) => {
-                return getFillOpacity(
-                    dataPoint.selected,
-                    dataPoint.highlight,
-                    !dataPoint.highlight && hasSelection,
-                    !dataPoint.selected && hasHighlights
-                );
-            });
-        }
-
-        private markDataPointsAsSelected(root: SunburstDataPoint): void {
-            if (!root || !root.parent) {
-                return;
+            if (onSelect) {
+                onSelect(dataPoint);
             }
+        });
 
-            root.parent.selected = true;
+        clearCatcher.on("click", () => {
+            selectionHandler.handleClearSelection();
 
-            this.markDataPointsAsSelected(root.parent);
-        }
+            if (onSelect) {
+                onSelect(null);
+            }
+        });
     }
 
-    export class InteractivityService extends InteractivityServiceBase {
-        constructor(host: IVisualHost, private onSelect?: (dataPoint: SunburstDataPoint) => void) {
-            super(host);
+    public renderSelection(hasSelection: boolean): void {
+        const {
+            selection,
+            interactivityService,
+        } = this.options;
+
+        this.options.dataPoints
+            .filter((dataPoint: SunburstDataPoint) => dataPoint && dataPoint.selected)
+            .forEach(this.markDataPointsAsSelected.bind(this));
+
+        const hasHighlights: boolean = interactivityService.hasSelection();
+
+        selection.style("opacity", (dataPoint: SunburstDataPoint) => {
+            return getFillOpacity(
+                dataPoint.selected,
+                dataPoint.highlight,
+                !dataPoint.highlight && hasSelection,
+                !dataPoint.selected && hasHighlights
+            );
+        });
+    }
+
+    private markDataPointsAsSelected(root: SunburstDataPoint): void {
+        if (!root || !root.parent) {
+            return;
         }
 
-        /**
-         * Sunburst does not support multi selection because it's hard to render a center tooltip for more than a single data point
-         */
-        public restoreSelection(selectionIds: visuals.ISelectionId[]): void {
-            super.restoreSelection(selectionIds);
+        root.parent.selected = true;
 
-            const selectedDataPoint: SunburstDataPoint = (this.selectableDataPoints as SunburstDataPoint[])
-                .filter((dataPoint: SunburstDataPoint) => {
-                    return dataPoint
-                        && dataPoint.identity
-                        && selectionIds
-                        && selectionIds[0]
-                        && selectionIds[0].equals(dataPoint.identity as visuals.ISelectionId);
-                })[0];
+        this.markDataPointsAsSelected(root.parent);
+    }
+}
 
-            if (this.onSelect) {
-                this.onSelect(selectedDataPoint);
-            }
+export class InteractivityService extends InteractivityServiceBase {
+    constructor(host: IVisualHost, private onSelect?: (dataPoint: SunburstDataPoint) => void) {
+        super(host);
+    }
+
+    /**
+     * Sunburst does not support multi selection because it's hard to render a center tooltip for more than a single data point
+     */
+    public restoreSelection(selectionIds: ISelectionId[]): void {
+        super.restoreSelection(selectionIds);
+
+        const selectedDataPoint: SunburstDataPoint = (this.selectableDataPoints as SunburstDataPoint[])
+            .filter((dataPoint: SunburstDataPoint) => {
+                return dataPoint
+                    && dataPoint.identity
+                    && selectionIds
+                    && selectionIds[0]
+                    && selectionIds[0].equals(dataPoint.identity as ISelectionId);
+            })[0];
+
+        if (this.onSelect) {
+            this.onSelect(selectedDataPoint);
         }
     }
 }
