+++ tsconfig.json
@@ -1,28 +1,16 @@
 {
-  "compilerOptions": {
-    "allowJs": false,
-    "alwaysStrict": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "target": "ES5",
-    "sourceMap": true,
-    "out": "./.tmp/build/visual.js",
-    "declaration": true
-  },
-  "files": [
-    ".api/v1.13.0/PowerBI-visuals.d.ts",
-    "node_modules/powerbi-models/dist/models-noexports.d.ts",
-    "node_modules/powerbi-visuals-utils-dataviewutils/lib/index.d.ts",
-    "node_modules/powerbi-visuals-utils-tooltiputils/lib/index.d.ts",
-    "node_modules/powerbi-visuals-utils-typeutils/lib/index.d.ts",
-    "node_modules/powerbi-visuals-utils-svgutils/lib/index.d.ts",
-    "node_modules/powerbi-visuals-utils-formattingutils/lib/index.d.ts",
-    "node_modules/powerbi-visuals-utils-colorutils/lib/index.d.ts",
-    "node_modules/powerbi-visuals-utils-chartutils/lib/index.d.ts",
-    "node_modules/powerbi-visuals-utils-interactivityutils/lib/index.d.ts",
-    "src/dataInterfaces.ts",
-    "src/settings.ts",
-    "src/behavior.ts",
-    "src/sunburst.ts"
-  ]
-}
+    "compilerOptions": {
+      "allowJs": false,
+      "emitDecoratorMetadata": true,
+      "experimentalDecorators": true,
+      "module": "es6",
+      "target": "ES6",
+      "sourceMap": true,
+      "outDir": "./.tmp/build/",
+      "moduleResolution": "node",
+      "declaration": true
+    },
+    "files": [
+      "./src/visual.ts"
+    ]
+  }
